Add support to ugly_url and local_markdown_link
If ugly_url is set to true in config.toml, then we generate urls ending
in .html. For example, page.md will  generate page.html.

If local_markdown_link is set to true in config.toml, then common
markdown links to local .md files will be parsed as local links and be
converted to .html links after checking that the file exists.  This
supports relative links.

It also makes page.meta.path ending in .html generate .html files
instead of a directory ending in .html. If the meta.path ending in /,
then it generates a path/index.html. Otherwise it behaves as specified
by ugly_url.